+++ rend3/src/util/output.rs
@@ -1,13 +1,10 @@
 use std::sync::Arc;
-use wgpu::{SurfaceError, SurfaceFrame, TextureView, TextureViewDescriptor};
+use wgpu::{SurfaceError, SurfaceTexture, TextureView, TextureViewDescriptor};
 
 use crate::types::Surface;
 
 pub enum OutputFrame {
-    Surface {
-        view: TextureView,
-        surface: Arc<SurfaceFrame>,
-    },
+    Surface { view: TextureView, surface: SurfaceTexture },
     View(Arc<TextureView>),
 }
 
@@ -17,7 +14,7 @@ impl OutputFrame {
         let mut retrieved_frame = None;
         for _ in 0..10 {
             profiling::scope!("Inner Acquire Loop");
-            match surface.get_current_frame() {
+            match surface.get_current_texture() {
                 Ok(frame) => {
                     retrieved_frame = Some(frame);
                     break;
@@ -28,18 +25,24 @@ impl OutputFrame {
         }
         let frame = retrieved_frame.expect("Swapchain acquire timed out 10 times.");
 
-        let view = frame.output.texture.create_view(&TextureViewDescriptor::default());
+        let view = frame.texture.create_view(&TextureViewDescriptor::default());
 
-        Ok(OutputFrame::Surface {
-            surface: Arc::new(frame),
-            view,
-        })
+        Ok(OutputFrame::Surface { surface: frame, view })
     }
 
     pub fn as_view(&self) -> &TextureView {
         match self {
             Self::Surface { view, .. } => view,
             Self::View(inner) => &**inner,
+        }
+    }
+
+    pub fn present(self) {
+        match self {
+            Self::Surface { surface, .. } => {
+                surface.present();
+            }
+            _ => {}
         }
     }
 }